Moved Compare, ModeContext, and StyleInProgress to local components
