+++ sandbox/grommet-app/src/pages/sticker-sheet/components/ModeContext.jsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const ModeContext = React.createContext({});
