+++ sandbox/grommet-app/src/pages/sticker-sheet/components/StyleInProgress.jsx
@@ -1,0 +1,9 @@
+import { Notification } from 'grommet';
+
+export const StyleInProgress = () => (
+  <Notification
+    status="warning"
+    message="This component style is still being refined."
+    margin={{ bottom: 'medium' }}
+  />
+);
