+++ sandbox/grommet-app/src/pages/sticker-sheet/components/index.js
@@ -1,0 +1,3 @@
+export * from './Compare';
+export * from './ModeContext';
+export * from './StyleInProgress';
